pyproject.toml
[tool.poetry]
name = "django-blog"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=4.0,<5.0"
django-environ = "^0.11.2"
crispy-bootstrap5 = "^2024.2"
django-taggit = "^5.0.1"
pillow = "^10.3.0"
social-auth-app-django = "^5.4.1"
django-tinymce = "^4.0.0"
django-htmx = "^1.17.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.2"
django-debug-toolbar = "^4.4.2"
ipython = "^8.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
